refactor(GlassHeroScene): remove unused code and types to streamline component, enhancing readability and maintainability
