+++ src/components/glass-hero/hooks/usePrefersReducedMotion.ts
@@ -1,0 +1,15 @@
+import { useSyncExternalStore } from "react";
+
+function subscribe(cb: () => void) {
+  const mq = window.matchMedia("(prefers-reduced-motion: reduce)");
+  mq.addEventListener("change", cb);
+  return () => mq.removeEventListener("change", cb);
+}
+
+function getSnapshot() {
+  return window.matchMedia("(prefers-reduced-motion: reduce)").matches;
+}
+
+export function usePrefersReducedMotion() {
+  return useSyncExternalStore(subscribe, getSnapshot, () => false);
+}
